git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@694 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/main/java/cspfj/filter/AC3.java
@@ -32,7 +32,7 @@ public final class AC3 implements Filter {
     public AC3(final Problem problem) {
         this(problem, new BinomialHeap<Variable>(new Key<Variable>() {
             @Override
-            public float getKey(final Variable object) {
+            public double getKey(final Variable object) {
                 return object.dom().size();
             }
         }));
