+++ src/test/java/cspfj/priorityqueues/BinomialHeapTest.java
@@ -29,7 +29,7 @@ public final class BinomialHeapTest {
                 new Key<IdInteger>() {
 
                     @Override
-                    public float getKey(final IdInteger object) {
+                    public double getKey(final IdInteger object) {
                         return object.value;
                     }
