+++ src/test/java/cspfj/priorityqueues/FifoTest.java
@@ -42,7 +42,7 @@ public final class FifoTest {
         final Fifos<IdInteger> fifo = new Fifos<IdInteger>(new Key<IdInteger>() {
 
             @Override
-            public float getKey(final IdInteger object) {
+            public double getKey(final IdInteger object) {
                 return object.value;
             }
         });
