+++ src/test/java/cspfj/priorityqueues/SoftHeapTest.java
@@ -33,7 +33,7 @@ public class SoftHeapTest {
                 new Key<IdInteger>() {
 
                     @Override
-                    public float getKey(final IdInteger object) {
+                    public double getKey(final IdInteger object) {
                         return object.value;
                     }
